Monorepo: introduce reporting mode for perf utility.
