+++ tools/compare-perf/package.json
@@ -7,15 +7,19 @@
 	"license": "GPLv2",
 	"repository": "woocommerce/woocommerce",
 	"scripts": {
-		"compare": "node index.js",
+		"compare": "node -r ts-node/register index.js",
 		"log": "node log-to-codevitals.js"
 	},
 	"dependencies": {
+		"@types/node": "^16.18.68",
+		"@tsconfig/node16": "^1.0.4",
 		"@wordpress/env": "^10.1.0",
 		"commander": "9.5.0",
 		"chalk": "^4.1.2",
 		"inquirer": "^7.1.0",
-		"simple-git": "^3.21.0"
+		"simple-git": "^3.21.0",
+		"ts-node": "^10.9.2",
+		"tslib": "^2.6.2"
 	},
 	"engines": {
 		"node": "^20.11.1",
