+++ tools/compare-perf/tsconfig.json
@@ -1,0 +1,15 @@
+{
+	"extends": "./node_modules/@tsconfig/node16/tsconfig.json",
+	"compilerOptions": {
+		"module": "Node16",
+		"moduleResolution": "Node16",
+		"typeRoots": [
+			"./typings",
+			"./node_modules/@types"
+		]
+	},
+	"ts-node": {
+		"transpileOnly": true,
+		"files": true,
+	},
+}
